Added tests for dns resolver
